mlservices/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
requests
beautifulsoup4